Move trace writes off provider path
